requirements.txt
asgiref==3.3.4
boto3==1.17.69
botocore==1.20.69
cffi==1.14.5
cryptography==3.4.7
Django==3.2.2
django-filter==2.4.0
django-rest-knox==4.1.0
django-storages==1.11.1
djangorestframework==3.12.4
gunicorn==20.1.0
jmespath==0.10.0
Pillow==8.2.0
pycparser==2.20
python-dateutil==2.8.1
pytz==2021.1
s3transfer==0.4.2
six==1.16.0
sqlparse==0.4.1
urllib3==1.26.4
whitenoise==5.2.0
